requirements.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
annotated-types==0.7.0
anyio==4.9.0
builtin-interfaces==2.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
composition-interfaces==2.0.2
contourpy==1.3.2
control-msgs==5.4.0
controller-manager==4.28.1
controller-manager-msgs==4.28.1
cryptography==44.0.3
cv-bridge==4.1.0
cycler==0.12.1
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
diagnostic-updater==4.2.3
domain-coordinator==0.12.0
ecdsa==0.19.1
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
fastapi==0.115.12
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
generate-parameter-library-py==0.4.0
geometry-msgs==5.3.6
gps-msgs==2.0.4
greenlet==3.2.1
h11==0.16.0
idna==3.10
image-geometry==4.1.0
interactive-markers==2.5.4
Jinja2==3.1.6
joy-teleop==1.8.0
joy-tester==0.0.2
kiwisolver==1.4.8
laser-geometry==2.7.0
launch==3.4.4
launch-ros==0.26.7
launch-testing==3.4.4
launch-testing-ros==0.26.7
launch-xml==3.4.4
launch-yaml==3.4.4
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
MarkupSafe==3.0.2
matplotlib==3.10.1
message-filters==4.11.6
mpmath==1.3.0
nav-msgs==5.3.6
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
osrf-pycommon==2.1.6
packaging==25.0
pal-statistics==2.6.2
pal-statistics-msgs==2.6.2
pandas==2.2.3
passlib==1.7.4
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
pillow==11.2.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
python-qt-binding==2.2.1
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
realsense2-camera-msgs==4.55.1
requests==2.32.3
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.11
ros-gz-interfaces==1.0.11
ros-gz-sim==1.0.11
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.7
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.2
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
rsa==4.9.1
scipy==1.15.2
seaborn==0.13.2
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.2
setuptools==80.3.1
shape-msgs==5.3.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sros2==0.13.3
starlette==0.46.2
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
sympy==1.14.0
teleop-tools-msgs==1.8.0
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.36.10
tf2-kdl==0.36.10
tf2-msgs==0.36.10
tf2-py==0.36.10
tf2-ros-py==0.36.10
tf2-sensor-msgs==0.36.10
tf2-tools==0.36.10
theora-image-transport==4.0.4
topic-monitor==0.33.5
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
trajectory-msgs==5.3.6
triton==3.3.0
turtle-tf2-py==0.5.0
turtlesim==1.8.3
type-description-interfaces==2.0.2
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.128
ultralytics-thop==2.0.14
unique-identifier-msgs==2.5.0
urllib3==2.4.0
uvicorn==0.34.2
vision-msgs==4.1.1
visualization-msgs==5.3.6
xacro==2.0.13
